requirements.txt
torch
datasets
accelerate==0.25.0
bitsandbytes
faiss-cpu
sentence-transformers
rank-bm25
peft==0.7.1
pdfplumber
pdf2image
pytesseract
nltk
numpy
scikit-learn
streamlit
huggingface_hub
optimum
python-dotenv
pyarrow
transformers==4.36.2
ctransformers>=0.2.27
adapter-transformers
huggingface_hub

